+++ packages/data-context/src/sources/MigrationDataSource.ts
@@ -96,35 +96,6 @@ export class MigrationDataSource {
     return this.ctx.lifecycleManager.metaState.needsCypressJsonMigration && Boolean(legacyConfigFileExists)
   }
 
-  async getVideoEmbedHtml () {
-    if (this.ctx.coreData.migration.videoEmbedHtml) {
-      return this.ctx.coreData.migration.videoEmbedHtml
-    }
-
-    const versionData = await this.ctx.versions.versionData()
-    const embedOnLink = `https://on.cypress.io/v10-video-embed/${versionData.current.version}`
-
-    // Time out request if it takes longer than 3 seconds
-    const controller = new AbortController()
-    const timeoutId = setTimeout(() => controller.abort(), 3000)
-
-    try {
-      const response = await this.ctx.util.fetch(embedOnLink, { method: 'GET', signal: controller.signal })
-      const { videoHtml } = await response.json()
-
-      this.ctx.update((d) => {
-        d.migration.videoEmbedHtml = videoHtml
-      })
-
-      return videoHtml
-    } catch {
-      // fail silently, no user-facing error is needed
-      return null
-    } finally {
-      clearTimeout(timeoutId)
-    }
-  }
-
   async getComponentTestingMigrationStatus () {
     debug('getComponentTestingMigrationStatus: start')
     if (!this.legacyConfig || !this.ctx.currentProject) {