+++ packages/launchpad/cypress/e2e/open-mode.cy.ts
@@ -1,21 +1,30 @@
 import type { SinonStub } from 'sinon'
 import defaultMessages from '@packages/frontend-shared/src/locales/en-US.json'
+import { MAJOR_VERSION_FOR_CONTENT } from '@packages/types/src'
 
 describe('Launchpad: Open Mode', () => {
   describe('global mode', () => {
-    beforeEach(() => {
+    it('shows Add Project when no projects have been added', () => {
       cy.openGlobalMode()
       cy.visitLaunchpad()
-    })
-
-    it('shows Add Project when no projects have been added', () => {
+      cy.skipWelcome()
       cy.get('h1').should('contain', defaultMessages.globalPage.empty.title)
     })
 
     it('shows the projects page when a project is not specified', () => {
+      cy.openGlobalMode()
       cy.addProject('todos')
       cy.visitLaunchpad()
+      cy.skipWelcome()
       cy.contains(defaultMessages.globalPage.recentProjectsHeader)
+    })
+
+    it('does not show Welcome screen on next Launchpad visit when it has already been dismissed', () => {
+      cy.visitLaunchpad()
+      cy.skipWelcome()
+      cy.visitLaunchpad()
+      cy.contains('button', 'Continue').should('not.exist')
+      cy.get('h1').should('contain', defaultMessages.globalPage.empty.title)
     })
   })
 
@@ -23,6 +32,7 @@ describe('Launchpad: Open Mode', () => {
     cy.scaffoldProject('todos')
     cy.openProject('todos', ['--e2e'])
     cy.visitLaunchpad()
+    cy.skipWelcome()
     cy.get('[data-cy=header-bar-content]').contains('e2e testing', { matchCase: false })
     // e2e testing is configured for the todo project, so we don't expect an error.
     cy.get('h1').should('contain', 'Choose a browser')
@@ -40,6 +50,7 @@ describe('Launchpad: Open Mode', () => {
 
     it('includes x-framework and x-dev-server, even when launched in e2e mode', () => {
       cy.visitLaunchpad()
+      cy.skipWelcome()
       cy.get('h1').should('contain', 'Choose a browser')
       cy.withCtx((ctx, o) => {
         expect(ctx.util.fetch).to.have.been.calledWithMatch('https://download.cypress.io/desktop.json', {
@@ -54,6 +65,7 @@ describe('Launchpad: Open Mode', () => {
     describe('logged-in state', () => {
       it(`sends 'false' when not logged in`, () => {
         cy.visitLaunchpad()
+        cy.skipWelcome()
         cy.get('h1').should('contain', 'Choose a browser')
         cy.withCtx((ctx, o) => {
           expect(ctx.util.fetch).to.have.been.calledWithMatch('https://download.cypress.io/desktop.json', {
@@ -67,6 +79,7 @@ describe('Launchpad: Open Mode', () => {
       it(`sends 'true' when logged in`, () => {
         cy.loginUser()
         cy.visitLaunchpad()
+        cy.skipWelcome()
         cy.get('h1').should('contain', 'Choose a browser')
         cy.withCtx((ctx, o) => {
           expect(ctx.util.fetch).to.have.been.calledWithMatch('https://download.cypress.io/desktop.json', {
@@ -83,6 +96,7 @@ describe('Launchpad: Open Mode', () => {
     cy.scaffoldProject('launchpad')
     cy.openProject('launchpad', ['--component'])
     cy.visitLaunchpad()
+    cy.skipWelcome()
     cy.get('[data-cy=header-bar-content]').contains('component testing', { matchCase: false })
     // Component testing is not configured for the todo project
     cy.get('h1').should('contain', 'Project setup')
@@ -102,13 +116,14 @@ describe('Launchpad: Open Mode', () => {
 
       // Need to visit after args have been configured, todo: fix in #18776
       cy.visitLaunchpad()
+      cy.skipWelcome()
       cy.contains('E2E Testing').click()
       cy.get('h1').should('contain', 'Choose a browser')
       cy.get('[data-cy-browser=firefox]').should('have.attr', 'aria-checked', 'true')
       cy.get('button[data-cy=launch-button]').invoke('text').should('include', 'Start E2E Testing in Firefox')
     })
 
-    it('auto-launches the browser when launched with --browser --testingType --project', () => {
+    it('auto-launches the browser when launched with --browser --testingType --project, after Major Version Welcome is dismissed', () => {
       cy.scaffoldProject('launchpad')
       cy.openProject('launchpad', ['--browser', 'firefox', '--e2e'])
       cy.withCtx((ctx, o) => {
@@ -117,6 +132,33 @@ describe('Launchpad: Open Mode', () => {
 
       // Need to visit after args have been configured, todo: fix in #18776
       cy.visitLaunchpad()
+
+      cy.skipWelcome()
+      cy.get('h1').should('contain', 'Choose a browser')
+      cy.get('[data-cy-browser=firefox]').should('have.attr', 'aria-checked', 'true')
+      cy.get('button[data-cy=launch-button]').invoke('text').should('include', 'Start E2E Testing in Firefox')
+
+      cy.withRetryableCtx((ctx) => {
+        expect(ctx._apis.projectApi.launchProject).to.be.calledOnce
+      })
+    })
+
+    it('auto-launches the browser when launched with --browser --testingType --project if there is no major version welcome screen', () => {
+      cy.withCtx((ctx, o) => {
+        o.sinon.stub(ctx._apis.projectApi, 'launchProject').resolves()
+        o.sinon.stub(ctx._apis.localSettingsApi, 'getPreferences').resolves({ majorVersionWelcomeDismissed: {
+          [o.MAJOR_VERSION_FOR_CONTENT]: Date.now(),
+        } })
+      }, {
+        MAJOR_VERSION_FOR_CONTENT,
+      })
+
+      cy.scaffoldProject('launchpad')
+      cy.openProject('launchpad', ['--browser', 'firefox', '--e2e'])
+
+      // Need to visit after args have been configured, todo: fix in #18776
+      cy.visitLaunchpad()
+
       cy.get('h1').should('contain', 'Choose a browser')
       cy.get('[data-cy-browser=firefox]').should('have.attr', 'aria-checked', 'true')
       cy.get('button[data-cy=launch-button]').invoke('text').should('include', 'Start E2E Testing in Firefox')
@@ -132,6 +174,7 @@ describe('Launchpad: Open Mode', () => {
       cy.scaffoldProject('todos')
       cy.openProject('todos')
       cy.visitLaunchpad()
+      cy.skipWelcome()
 
       cy.withCtx(async (ctx, o) => {
         ctx.emitter.toLaunchpad()
@@ -146,6 +189,7 @@ describe('Launchpad: Open Mode', () => {
       cy.scaffoldProject('todos')
       cy.openProject('todos')
       cy.visitLaunchpad()
+      cy.skipWelcome()
 
       cy.contains('button', 'Docs').click()
       cy.contains(defaultMessages.topNav.docsMenu.gettingStartedTitle).should('be.visible')
@@ -175,6 +219,7 @@ describe('Launchpad: Open Mode', () => {
       })
 
       cy.visitLaunchpad()
+      cy.skipWelcome()
       cy.findByTestId('project-card')
       cy.get('[aria-label="Project actions"]').click()
       cy.get('button').contains('Open in IDE').click()
@@ -201,6 +246,7 @@ describe('Launchpad: Open Mode', () => {
       })
 
       cy.visitLaunchpad()
+      cy.skipWelcome()
       cy.findByTestId('project-card')
       cy.get('[aria-label="Project actions"]').click()
 
@@ -219,6 +265,7 @@ describe('Launchpad: Open Mode', () => {
     cy.scaffoldProject('no-support-file')
     cy.openProject('no-support-file', ['--e2e'])
     cy.visitLaunchpad()
+    cy.skipWelcome()
     cy.contains(cy.i18n.launchpadErrors.generic.configErrorTitle)
     cy.contains('Your project does not contain a default supportFile.')
     cy.contains('If a support file is not necessary for your project, set supportFile to false.')
@@ -231,6 +278,7 @@ describe('Launchpad: Open Mode', () => {
     cy.scaffoldProject('project-with-(glob)-[chars]')
     cy.openProject('project-with-(glob)-[chars]', ['--e2e'])
     cy.visitLaunchpad()
+    cy.skipWelcome()
 
     cy.get('body').should('not.contain.text', 'Your project does not contain a default supportFile.')
     cy.get('h1').should('contain', 'Choose a browser')
@@ -240,6 +288,7 @@ describe('Launchpad: Open Mode', () => {
     cy.scaffoldProject('simple with spaces')
     cy.openProject('simple with spaces', ['--e2e'])
     cy.visitLaunchpad()
+    cy.skipWelcome()
     cy.get('h1').should('contain', 'Choose a browser')
   })
 })