+++ scripts/verify-mocha-results.js
@@ -95,8 +95,6 @@ async function checkReportFiles (filenames) {
 }
 
 async function verifyMochaResults () {
-  if (process.platform === 'win32') throw new Error('verifyMochaResults not supported on Windows')
-
   try {
     const filenames = await fs.readdir(REPORTS_PATH)