+++ system-tests/projects/v8-snapshot/require-full-path-var/app.js
@@ -1,0 +1,7 @@
+'use strict'
+// Roughly replicates https://github.com/kribblo/node-ffmpeg-installer/blob/master/index.js
+
+const getPackage = require('./entry')
+const pack = getPackage()
+
+console.log(JSON.stringify({ version: pack.version }))